Refactor: Extract resource watching logic from drpc controller into drpc watcher
- Moved all resource-watching related functions from drplacementcontrol_controller.go
to a new file drplacementcontrol_watcher.go.
- No functional changes introduced, purely a structural refactor.
